arm: DT: Scorpion: Refactor panels to allow video mode and PCC calibration
Scorpion's panels implementation was referring to cmd mode panels,
but we don't want to always use cmd for various reasons, such as
performance degradation and driver problems.

Refactor the panel timings and initialization to allow video mode
to be used with those ones, also solving the pixels corruption
issue happening on the current kernel revision as of 08/05/2015
(BF64.1.2.1-RB1.05.00.02.019.067).

Also, add a flag to force PCC calibration, because the panel will
not give a good reply to our UV commands, and we still need to
calibrate it. This will add no performance overhead.