video: msm: MHL: Add SII8620 complete cleaned up driver for Shinano
The previous driver we included in the kernel was incomplete.
It was required to allow the charger driver to be happy by
creating the required MHL power supply.

Also, that driver was coming from the latest release for Leo and
that was incomplete.

The new driver comes from an old release (unfortunately), intended
for Sirius.