video: msm: mdss: Support for ULP modes. Code style cleanup.
Modify display power on function to support multiple modes and
to use the enumeration for checking the new power mode.
While at it, add the function for LP and ULP modes (LP1/LP2)
for the displays that will support those in the future.

Also, cleanup code style where possible without having impacts
on human readability.