yas / kxtj2: put back some old code for gpio determination.
This was previously in the driver, left out on a sleep fix, but was only tested on one hardwareversion of D2303.

This should return a gpio validation that is lost on current code.

on my D2303 the gpio is 65

According to http://dl-1.va.us.xda-developers.com/3/2/7/0/7/5/9/zipkernel.txt?key=dxh0C8bCbj1DeMxcEWHBCg&ts=1434569576
the gio on the alternative seems to be 64.